+++ tests/testthat/test-plotting.R
@@ -3,17 +3,21 @@ context("Plotting")
 test_that("normal plotting works without errors", {
   f1 <- euler(c("A" = 10, "B" = 5, "A&B" = 2))
 
-  expect_silent(dont_plot(f1,
-                          fills = list(fill = c("black", "blue"), alpha = 0.3),
-                          labels = list(font = 4)))
+  p <- list()
 
-  expect_silent(dont_plot(f1))
-  expect_silent(dont_plot(f1, legend = TRUE, quantities = TRUE))
-  expect_silent(dont_plot(f1, legend = FALSE, quantities = TRUE))
-  expect_silent(dont_plot(f1, legend = TRUE, labels = FALSE))
-  expect_silent(dont_plot(f1, labels = c("asdf", "qwer")))
-  expect_silent(dont_plot(f1, main = "Hello"))
-  expect_silent(dont_plot(f1, expression = "phi[1]"))
+  expect_silent(p[[1]] <- plot(f1,
+                               fills = list(fill = c("black", "blue"),
+                                            alpha = 0.3),
+                               labels = list(font = 4)))
+
+  expect_silent(p[[2]] <- plot(f1))
+  expect_silent(p[[3]] <- plot(f1, legend = TRUE, quantities = TRUE))
+  expect_silent(p[[4]] <- plot(f1, legend = FALSE, quantities = TRUE))
+  expect_silent(p[[5]] <- plot(f1, legend = TRUE, labels = FALSE))
+  expect_silent(p[[6]] <- plot(f1, labels = c("asdf", "qwer")))
+  expect_silent(p[[7]] <- plot(f1, main = "Hello"))
+  expect_silent(p[[8]] <- plot(f1, expression = "phi[1]"))
+  expect_silent(p[[9]] <- plot(f1, edges = c("white", "blue")))
 
   grid <- expand.grid(labels = c(TRUE, FALSE),
                       quantities = c(TRUE, FALSE),
@@ -22,14 +26,13 @@ test_that("normal plotting works without errors", {
                       edges = c(TRUE, FALSE))
 
   for (i in seq_len(nrow(grid))) {
-    expect_silent(dont_plot(f1,
-                            labels = grid$labels[i],
-                            quantities = grid$quantities[i],
-                            legend = grid$legend[i],
-                            fills = grid$fills[i],
-                            edges = grid$edges[i]))
+    expect_silent(dont_plot(plot(f1,
+                                 labels = grid$labels[!!i],
+                                 quantities = grid$quantities[!!i],
+                                 legend = grid$legend[!!i],
+                                 fills = grid$fills[!!i],
+                                 edges = grid$edges[!!i])))
   }
-
 
   dat <- data.frame(
     Liberal = sample(c(TRUE, FALSE), size = 100, replace = TRUE),
@@ -41,10 +44,10 @@ test_that("normal plotting works without errors", {
   f2 <- euler(dat[, 1:3], by = list(Gender))
   f3 <- euler(dat, by = list(Gender, Nation))
 
-  expect_silent(dont_plot(f2,
-                          fills = list(fill = "transparent",
-                                       lty = c(1, 2),
-                                       lwd = c(1, 2))))
+  expect_silent(p[[10]] <- plot(f2,
+                                fills = list(fill = "transparent",
+                                             lty = c(1, 2),
+                                             lwd = c(1, 2))))
 
   grid <- expand.grid(side = c(NA, "right", "left", "top", "bottom"),
                       main = c("Title", FALSE))
@@ -55,18 +58,17 @@ test_that("normal plotting works without errors", {
     } else {
       legend <- list(side = grid$side[i])
     }
-    main <- grid$main[i]
 
-    expect_silent(dont_plot(f1, legend = legend, main = main))
-    expect_silent(dont_plot(f2, legend = legend, main = main))
-    expect_silent(dont_plot(f3, legend = legend, main = main))
+    expect_silent(dont_plot(plot(f1, legend = legend, main = grid$main[!!i])))
+    expect_silent(dont_plot(plot(f2, legend = legend, main = grid$main[!!i])))
+    expect_silent(dont_plot(plot(f3, legend = legend, main = grid$main[!!i])))
   }
 
-  expect_error(euler(dat, by = list(Gender, Nation, Gender)))
-
-
   f4 <- euler(c(A = 1))
-  expect_silent(dont_plot(f4))
+  expect_silent(p[[11]] <- plot(f4))
+
+  for (i in seq_along(p))
+    expect_silent(dont_plot(p[[!!i]]))
 })
 
 test_that("plotting zero-fits works", {
@@ -77,11 +79,18 @@ test_that("plotting zero-fits works", {
 test_that("error_plot functions normally", {
   f <- euler(organisms)
 
-  expect_silent(dont_plot(f, .f = error_plot))
-  expect_silent(dont_plot(f,
-                          pal = grDevices::colorRampPalette(c("red", "blue")),
-                          .f = error_plot))
-  expect_silent(dont_plot(f, quantities = FALSE, .f = error_plot))
-  expect_silent(dont_plot(f, edges = FALSE, .f = error_plot))
+  expect_silent(dont_plot(error_plot(f)))
+  expect_silent(dont_plot(error_plot(f,
+                                     pal = grDevices::colorRampPalette(c("red", "blue")))))
+  expect_silent(dont_plot(error_plot(f, quantities = FALSE)))
+  expect_silent(dont_plot(error_plot(f, edges = FALSE)))
 })
 
+test_that("plots with euler lists works", {
+  f1 <- euler(fruits[, 1:5], by = age)
+  f2 <- euler(fruits[, 1:5], by = list(age, sex))
+
+  expect_silent(dont_plot(plot(f1, legend = TRUE, strips = FALSE)))
+  expect_silent(dont_plot(plot(f2, strips = list(cex = 2, fontface = "bold"))))
+  expect_silent(dont_plot(plot(f1)))
+})