+++ tests/testthat/test-utils.R
@@ -1,5 +1,5 @@
 context("Utlities and misc")
-#
+
 test_that("getting options works", {
   expect_type(eulerr_options(), "list")
 })